requirements.txt
# Core
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
PyYAML>=6.0
requests>=2.31.0

# VLM: InternVL-2.5 (local)
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
accelerate>=0.26.0
sentencepiece>=0.1.99
einops>=0.7.0
timm>=0.9.0

# Optional VLM backends
openai>=1.10.0              # GPT-4o
google-generativeai>=0.4.0  # Gemini-1.5-Flash

# Preprocessing (optional but recommended)
lang-sam>=0.1.0             # Lang-SAM (installs GroundingDINO + SAM2 deps)
opencv-python>=4.8.0
scipy>=1.11.0

# Dataset download
dataset-tools>=0.1.0

# Evaluation & logging
matplotlib>=3.7.0
pandas>=2.0.0
